requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
Deprecated==1.2.13
dimod==0.12.3
diskcache==5.4.0
dwave-cloud-client==0.10.4
dwave-greedy==0.3.0
dwave-hybrid==0.6.10
dwave-inspector==0.4.1
dwave-neal==0.6.0
dwave-networkx==0.8.13
dwave-ocean-sdk==6.3.0
dwave-preprocessing==0.5.4
dwave-samplers==1.0.0
dwave-system==1.18.0
dwave-tabu==0.5.0
dwavebinarycsp==0.2.0
fasteners==0.18
Flask==2.2.2
homebase==1.0.1
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.10.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
minorminer==0.2.11
morejson==1.1.5
networkx==2.8.8
numpy==1.24.2
packaging==23.0
pandas==1.5.3
patsy==0.5.3
penaltymodel==1.0.2
plucky==0.4.3
PuLP==2.7.0
pydantic==1.10.4
pyqubo==1.4.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.7.1
rectangle-packer==2.0.1
requests==2.28.2
scipy==1.10.0
six==1.16.0
statsmodels==0.13.5
typing-extensions==4.4.0
urllib3==1.26.14
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.13.0
